feat: repository initializing for tests

+++ src/Lab2/Entities/Hdd.cs
@@ -1,8 +1,10 @@
+using Itmo.ObjectOrientedProgramming.Lab2.Models;
+
 namespace Itmo.ObjectOrientedProgramming.Lab2.Entities;
 
 public class Hdd : IReposirotyAdded
 {
-    public Hdd(int capacity, int rotatingSpeed, int powerConsumption)
+    public Hdd(int capacity, int rotatingSpeed, double powerConsumption)
     {
         Capacity = capacity;
         RotatingSpeed = rotatingSpeed;
@@ -11,7 +13,7 @@ public class Hdd : IReposirotyAdded
 
     public int Capacity { get; }
     public int RotatingSpeed { get; }
-    public int PowerConsumption { get; }
+    public double PowerConsumption { get; }
 
     public void AddToRepository(Repository repository)
     {
