+++ src/Lab2/Entities/Ssd.cs
@@ -4,7 +4,7 @@ namespace Itmo.ObjectOrientedProgramming.Lab2.Entities;
 
 public class Ssd : IReposirotyAdded
 {
-    public Ssd(VariantConnectingSsd? connecting, int capacity, int maxSpeed, int powerConsumption)
+    public Ssd(VariantConnectingSsd? connecting, int capacity, int maxSpeed, double powerConsumption)
     {
         Connecting = connecting;
         Capacity = capacity;
@@ -15,7 +15,7 @@ public class Ssd : IReposirotyAdded
     public VariantConnectingSsd? Connecting { get; set; }
     public int Capacity { get; set; }
     public int MaxSpeed { get; set; }
-    public int PowerConsumption { get; set; }
+    public double PowerConsumption { get; set; }
 
     public void AddToRepository(Repository repository)
     {
